relocate EventHistoryEntry to appropriate location, export from index
