+++ src/api/mungers.ts
@@ -89,28 +89,3 @@ export function mungeSimBundleOptions(params: SimBundleOptions) {
         baseFee: params.baseFee && `0x${params.baseFee.toString(16)}`,
     }
 }
-
-export class EventHistoryEntry {
-    public block: number
-    public timestamp: number
-    public hint: {
-        txs?: Array<{
-            to: string,
-            callData: string,
-            functionSelector: string,
-        }>,
-        hash: string,
-        logs?: Array<LogParams>,
-        gasUsed: bigint,
-        mevGasPrice: bigint,
-    }
-    constructor(entry: IEventHistoryEntry) {
-        this.block = entry.block
-        this.timestamp = entry.timestamp
-        this.hint = {
-            ...entry.hint,
-            gasUsed: BigInt(entry.hint.gasUsed),
-            mevGasPrice: BigInt(entry.hint.mevGasPrice),
-        }
-    }
-}
